[walnut] move LoadWorkflow into loadDPL.go
